Loading a selection from an osm file
A capability to load a selection from an osm file has been added.  The file must be in the JOSM xml format.
 Implemented an XML parser.

get_data() added to api.py for convivnience.  It gets data either from OSM by city name or from an XML file by file name.  If the city name is not None, than data will be downloaded from OSM online.  If the file name is not None, the data will be loaded from the file.  If both are not None, the file will be ignored and the city name parameter prevails.  Returns a dictionary with the desired data or None if not found.

Added loading_from_osm_file notebook as a sample use.