Implemented get_geo_coordinates to convert from (0,1) to geo coordinates
Convert normalized coordinates (between 0 and 1) to lon and lat.   point_of_view[0] is relative distance from the beginning of the median to the intersection with the conflict zone.  The first parameter of the point_of_view tuple is a distance along the median of the guideway.   0 is a point at the beginning of the guideway.  1 is a point at the intersection of the guideway median and the conflict zone.

If the conflict zone is None (not specified), then 1 means the end of the guideway.

point_of_view[1] is position within the width of the guideway, where 0.5 is on the median, 0 on the left border and 1 on the right border.

    :param point_of_view: a tuple of floats between 0 and 1
    :param conflict_zone: conflict zone dictionary
    :param guideway_data: guideway dictionary
    :return: a tuple of lon and lat