Gudideway reduction by relative distance
Reduce guideway by relative distance from either end.  The distance is in the range [0;1].   The starting point can be either 'b' or 'e';  The guideway left and right borders and median will truncated.  For example, if relative_distance = 0.3 and starting_point_for_cut="b", then the function returns 30% of the original length starting from the  beginning of the guideway.  If relative_distance = 0.3 and starting_point_for_cut="e", then the function returns 30% of the original length adjacent to the end of the guideway.